Move MetadataTable utils to helpers
